Refactor: new ShowObject catalog hierarchy structure
