+++ mwdb/web/src/components/ShowObject/Actions/FavoriteAction.js
