+++ mwdb/web/src/components/ShowObject/Views/RelationsTab.js
@@ -4,7 +4,7 @@ import queryString from "query-string";
 
 import { faProjectDiagram, faSearch } from '@fortawesome/free-solid-svg-icons'
 
-import RelationsPlot from '../RelationsPlot';
+import RelationsPlot from '../../RelationsPlot';
 
 import { ObjectContext } from "@mwdb-web/commons/context";
 import { ObjectAction, ObjectTab } from "@mwdb-web/commons/ui";
